+++ src/ios/FacebookConnectPlugin.h
@@ -5,6 +5,7 @@
 //  Created by Jesse MacFadyen on 11-04-22.
 //  Updated by Ally Ogilvie on 29/Jan/2014.
 //  Updated by Jeduan Cornejo on 3/Jul/2015
+//  Updated by David Dal Busco on 21/Apr/2018 - Facebook doesn't support App Invites anymore
 //  Copyright 2011 Nitobi. All rights reserved.
 //
 
@@ -15,7 +16,7 @@
 #import <Cordova/CDV.h>
 #import "AppDelegate.h"
 
-@interface FacebookConnectPlugin : CDVPlugin <FBSDKSharingDelegate, FBSDKAppInviteDialogDelegate, FBSDKGameRequestDialogDelegate>
+@interface FacebookConnectPlugin : CDVPlugin <FBSDKSharingDelegate, FBSDKGameRequestDialogDelegate>
 - (void)getLoginStatus:(CDVInvokedUrlCommand *)command;
 - (void)getAccessToken:(CDVInvokedUrlCommand *)command;
 - (void)logEvent:(CDVInvokedUrlCommand *)command;
@@ -25,7 +26,6 @@
 - (void)logout:(CDVInvokedUrlCommand *)command;
 - (void)graphApi:(CDVInvokedUrlCommand *)command;
 - (void)showDialog:(CDVInvokedUrlCommand *)command;
-- (void)appInvite:(CDVInvokedUrlCommand *) command;
 - (void)getDeferredApplink:(CDVInvokedUrlCommand *) command;
 - (void)activateApp:(CDVInvokedUrlCommand *)command;
 @end